Move:player-resultをpagesフォルダに移動 #27
