+++ src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AddLeagueComponent } from './add-league/add-league.component';
 import { AddResultComponent } from './add-result/add-result.component';
-import { PlayerDetailsComponent } from './player-details/player-details.component';
+import { PlayerDetailsComponent } from './pages/player-details/player-details.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '', pathMatch: 'full' },
