+++ src/app/pages/player-details/player-details.component.scss
